ui/api: Split K8s APIs by domain
To help with organization of the increasing number of API call helpers,
we split them into domain-based modules, e.g. `k8s/solutions` or
`k8s/core`.
